+++ ui/src/routes/_layout.svelte
@@ -5,12 +5,12 @@
 	import {onMount, beforeUpdate, tick} from 'svelte';
 
 	import AccessibilityMenu from 'app/components/AccessibilityMenu.svelte';
-	import Beta from 'app/components/content/Beta.svelte';
+	import Beta from 'app/components/content/info/Beta.svelte';
 	import ColorCorrection from 'app/components/ColorCorrection.svelte';
 	import MultiBanner from 'app/components/MultiBanner.svelte';
 	import Nav from 'app/components/Nav.svelte';
 	import NoScript from 'app/components/NoScript.svelte';
-	import Privacy from 'app/components/content/Privacy.svelte';
+	import Privacy from 'app/components/content/info/Privacy.svelte';
 	import {
 		_a11yColorStyles,
 		_a11yTextStyles,