+++ ui/src/routes/indicators/[id]/[year].svelte
@@ -2,7 +2,7 @@
 	import {lookup} from 'app/data/groups';
 	import {parseCSV} from 'app/utils/domain';
 
-	export function preload({ params: {id, year}, query }) {
+	export function preload ({params: {id, year}}) {
 		return this.fetch(lookup[id].url)
 			.then(r => r.text())
 			.then(parseCSV(id))
@@ -40,7 +40,7 @@
 </script>
 
 <svelte:head>
-	<title>{toolName}: {title} ({year})</title>
+	<title>{title} ({year}) - {toolName}</title>
 	<meta
 		content='{toolName}: geographic distribution (NUTS2 regions) of the indicator: {title} ({year})'
 		name='description'