+++ ui/src/routes/info/disclaimer.svelte
@@ -1,3 +1,15 @@
+<script>
+	import {toolName} from 'app/config';
+</script>
+
+<svelte:head>
+	<title>Disclaimer - {toolName}</title>
+	<meta
+		content='{toolName}: disclaimer'
+		name='description'
+	>
+</svelte:head>
+
 <section>
 	<h2>Disclaimer</h2>
 	<p>