+++ ui/src/routes/info/privacy.svelte
@@ -1,5 +1,14 @@
 <script>
-	import Privacy from 'app/components/content/Privacy.svelte';
+	import Privacy from 'app/components/content/info/Privacy.svelte';
+	import {toolName} from 'app/config';
 </script>
+
+<svelte:head>
+	<title>Privacy - {toolName}</title>
+	<meta
+		content='{toolName}: Privacy statement'
+		name='description'
+	>
+</svelte:head>
 
 <Privacy />